.business/financieelplan_NL_nl/.003-draft/engine/requirements.txt
# D3 Engine (scaffold) runtime dependencies
# Prefer lean set; expand as needed during implementation
numpy>=1.26
pandas>=2.2
pydantic>=2.7
PyYAML>=6.0
pytest>=8.2
pytest-bdd>=7.0
psutil>=5.9
pytest-cov>=5.0
